% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{C15}
\alias{C15}
\title{C15}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 8048 rows and 2 columns.
}
\usage{
C15
}
\description{
Poligoni geografici dei comuni Italiani al 1 Gennaio 2015, in formato \code{sf} semplificati al 20%.
}
\details{
Il file contiene, oltre alle geometrie, i codici ISTAT dei Comuni (\code{CMN})
e corrispondenti Province (\code{PRV}), Regioni (\code{RGN}) e Ripartizioni (\code{RPT}). 
Maggiori informazioni possono essere reperite nella tabella \code{comuni} filtrata all'anno 2001.

Per ulteriori dettagli, consultare il sito ISTAT \url{https://www.istat.it/it/archivio/222527} e \url{https://www.istat.it/it/archivio/6789}.
}
\keyword{datasets}
