% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{comuni}
\alias{comuni}
\title{comuni}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{anno}}{ anno di riferimento }
  \item{\code{CMN}}{ Codice Istat }
  \item{\code{CMNd}}{ Descrizione }
  \item{\code{PRV}}{ Codice Istat PRovincia }
  \item{\code{RGN}}{ Codice Istat Regione }
  \item{\code{RPT}}{ Codice Istat Ripartizione }
  \item{\code{x_lon}}{ Longitudine del centroide geografico }
  \item{\code{y_lat}}{ Latitudine del centroide geografico }
  \item{\code{area}}{ Area }
  \item{\code{perimetro}}{ Perimetro }
}
}
\usage{
comuni
}
\description{
Elenco dei Comuni divisi per anno di esistenza
}
\keyword{datasets}
