% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_comuni.R
\name{estrai_comuni}
\alias{estrai_comuni}
\title{estrai_comuni}
\usage{
estrai_comuni(x, anno = 22, tp = "CMN", sup = FALSE, mappa = FALSE)
}
\arguments{
\item{x}{codice/i delle localita da estrarre}

\item{anno}{anno di riferimento}

\item{tp}{tipo delle localita a cui devono associarsi i codici forniti, uno fra: (C)}

\item{sup}{indica se devono essere restituiti: 
(\code{FALSE}) tutti i comuni legati ad \code{x} e \code{tp}
(\code{TRUE}) esattamente le localita indicate in \code{x} e tipo \code{tp}}
}
\value{
un oggetto sf o leaflet
}
\description{
Estrae i poligoni di uno o piu' comuni, e se richiesto ne visualizza la mappa
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
