% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{provincie}
\alias{provincie}
\title{provincie}
\format{
Una data.table con i campi seguenti:
\describe{
  \item{\code{anno}}{ anno di riferimento }
  \item{\code{PRV}}{ Codice Istat }
  \item{\code{PRVd}}{ Descrizione }
  \item{\code{RGN}}{ Codice Istat Regione }
  \item{\code{RGNd}}{ Descrizione Regione }
  \item{\code{RPT}}{ Codice Istat Ripartizione }
  \item{\code{RPTd}}{ Descrizione Ripartizione }
  \item{\code{x_lon}}{ Longitudine del centroide geografico }
  \item{\code{y_lat}}{ Latitudine del centroide geografico }
  \item{\code{area}}{ Area }
  \item{\code{perimetro}}{ Perimetro }
}
}
\usage{
provincie
}
\description{
Elenco delle Province esistenti lungo gli anni, con corrispondenti Regioni e Ripartizioni
}
\keyword{datasets}
