% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{PRV}
\alias{PRV}
\title{PRV}
\format{
An object of class \code{list} of length 22.
}
\usage{
PRV
}
\description{
Lista dei poligoni geografici delle Provincie Italiane dal 1 Gennaio 2001 al 1 Gennaio 2022, 
in formato \code{sf} semplificati al 20%.
}
\details{
Il file contiene, oltre alle geometrie, i codici ISTAT delle Province (\code{PRV})
e corrispondenti Regioni (\code{RGN}) e Ripartizioni (\code{RPT}). 
Maggiori informazioni possono essere reperite nella tabella \code{provincie} filtrata al corrispondente anno.

Per ulteriori dettagli, consultare il sito ISTAT \url{https://www.istat.it/it/archivio/222527} e \url{https://www.istat.it/it/archivio/6789}.
}
\keyword{datasets}
