% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dati.R
\docType{data}
\name{RPT}
\alias{RPT}
\title{RPT}
\format{
An object of class \code{list} of length 22.
}
\usage{
RPT
}
\description{
Lista dei poligoni geografici delle Ripartizioni Italiane dal 1 Gennaio 2001 al 1 Gennaio 2022, 
in formato \code{sf} semplificati al 20%.
}
\details{
Il file contiene, oltre alle geometrie, i corrispondenti codici ISTAT.

Per ulteriori dettagli, consultare il sito ISTAT \url{https://www.istat.it/it/archivio/222527} e \url{https://www.istat.it/it/archivio/6789}.
}
\keyword{datasets}
